Apply redesign to hackearth&opensource chars && change some d3 file names for clarity
