+++ build.js
@@ -8,7 +8,8 @@ let entries = {
   "side-navigation": "./static/js/src/side-navigation.js",
   "image-download": "./static/js/src/image-download.js",
   main: "./static/js/src/main.js",
-  "release-chart": "./static/js/src/release-chart.js",
+  "release-chart-manager": "./static/js/src/release-chart-manager.js",
+  "developer-chart": "./static/js/src/developer-chart.js",
   tabotronic: "./static/js/src/tabotronic.js",
   appliance: "./static/js/src/appliance.js",
   costCalculator: "./static/js/src/openstack/react/app.jsx",
