+++ static/js/src/developer-chart.js
@@ -1,35 +1,263 @@
-const hackerEarthData = [
-  { count: 66, label: "Ubuntu" },
-  { count: 61, label: "MS Windows" },
-  { count: 57, label: "MacOS" },
-  { count: 11, label: "CentOs" },
-  { count: 10, label: "Debian" },
-  { count: 9, label: "Fedora" },
-  { count: 8, label: "Arch Linux" },
-  { count: 4, label: "Solaris" },
-  { count: 2, label: "FreeBSD" },
-  { count: 1, label: "Deepin" },
-];
-const openSourceData = [
-  { count: 35.6, label: "Ubuntu" },
-  { count: 21.4, label: "Debian" },
-  { count: 19.5, label: "CentOS" },
-  { count: 16, label: "RHEL" },
-  { count: 14.6, label: "OpenSUSE" },
-  { count: 13.5, label: "SLES" },
-  { count: 11.1, label: "SELinux" },
-  { count: 7.5, label: "Rocky Linux" },
-  { count: 4.8, label: "NavyLinux" },
-];
+import { debounce } from "./utils/debounce.js";
+
+var hackerEarthData = {
+  tasks: [
+    { value: 66, taskName: "Ubuntu" },
+    { value: 61, taskName: "MS Windows" },
+    { value: 57, taskName: "MacOS" },
+    { value: 11, taskName: "CentOs" },
+    { value: 10, taskName: "Debian" },
+    { value: 9, taskName: "Fedora" },
+    { value: 8, taskName: "Arch Linux" },
+    { value: 4, taskName: "Solaris" },
+    { value: 2, taskName: "FreeBSD" },
+    { value: 1, taskName: "Deepin" },
+  ],
+  labels: [
+    "Ubuntu",
+    "MS Windows",
+    "MacOS",
+    "CentOs",
+    "Debian",
+    "Fedora",
+    "Arch Linux",
+    "Solaris",
+    "FreeBSD",
+    "Deepin",
+  ],
+};
+
+var openSourceData = {
+  tasks: [
+    { value: 35.6, taskName: "Ubuntu" },
+    { value: 21.4, taskName: "Debian" },
+    { value: 19.5, taskName: "CentOS" },
+    { value: 16, taskName: "RHEL" },
+    { value: 14.6, taskName: "OpenSUSE" },
+    { value: 13.5, taskName: "SLES" },
+    { value: 11.1, taskName: "SELinux" },
+    { value: 7.5, taskName: "Rocky Linux" },
+    { value: 4.8, taskName: "NavyLinux" },
+  ],
+  labels: [
+    "Ubuntu",
+    "Debian",
+    "CentOS",
+    "RHEL",
+    "OpenSUSE",
+    "SLES",
+    "SELinux",
+    "Rocky Linux",
+    "NavyLinux",
+  ],
+};
+
+// Global variables - used widely throughout the chart making process
+var margin = {
+  top: 24,
+  right: 40,
+  bottom: 30,
+  left: 0,
+};
+var gapSpacing = 32;
 
 /**
- * 
- * @param {*} data 
- * @param {*} id 
+ *
+ * @param {*} data
  */
 function sortData(data) {
-  return data = data.sort(function (a, b) {
+  return (data = data.sort(function (a, b) {
     return d3.ascending(a.count, b.count);
+  }));
+}
+
+/**
+ *
+ * @param {*} svg
+ * @param {Array} tasks
+ * @param {*} x
+ * @param {*} y
+ *
+ * Add bars to chart using supplied data
+ */
+function addBarsToChart(svg, tasks, x, y) {
+  const barClass = (d) =>
+    d.taskName === "Ubuntu" ? "chart__bar--orange" : "chart__bar--light-grey";
+  const barWidth = (d) => x(d.value) - x(0);
+  const barTransform = (d) => `translate(${x(0)},${y(d.taskName)})`;
+
+  svg
+    .selectAll(".chart")
+    .data(tasks, (d) => d.taskName + d.value)
+    .enter()
+    .append("rect")
+    .attr("class", barClass)
+    .attr("y", 0)
+    .attr("transform", barTransform)
+    .attr("height", y.bandwidth)
+    .attr("width", barWidth);
+}
+
+/**
+ *
+ * @param {*} svg
+ * @param {Int} height
+ * @param {Object} margin
+ * @param {*} xAxis
+ *
+ * Add x-axis group to SVG
+ */
+function addXAxis(svg, height, margin, xAxis) {
+  var xAxisGroup = svg
+    .append("g")
+    .attr("class", "x axis")
+    .attr("transform", "translate(0, " + -margin.top + ")")
+    .call(xAxis);
+  xAxisGroup.selectAll(".tick text").attr("dx", 3);
+}
+
+/**
+ *
+ * @param {*} svg
+ * @param {*} yAxis
+ *
+ * Append y-axis group to SVG
+ */
+function appendYAxisGroup(svg, yAxis) {
+  return svg.append("g").attr("class", "y axis").call(yAxis);
+}
+
+/**
+ *
+ * @param {*} yAxisGroup
+ * @param {Array} tasks
+ * Example of tasks array:
+ * [{taskName: "Task 1", value: 10}, {taskName: "Task 2", value: 20}, ...]
+ *
+ * Loops through the axis labels and calls the formatLabel
+ * function on them
+ */
+function formatYAxisLabels(yAxisGroup, tasks) {
+  yAxisGroup.selectAll(".tick text").each(function (d, i) {
+    d3.select(this).text("");
+    d3.select(this).append("tspan").text(d);
+    d3.select(this)
+      .append("tspan")
+      .text(tasks[i].value + "%")
+      .attr("style", "text-anchor: end;")
+      .attr("x", -gapSpacing);
+  });
+}
+
+/**
+ *
+ * @param {*} svg
+ * @param {*} yAxis
+ * @param {Array} tasks
+ * Example of tasks array:
+ * [{taskName: "Task 1", value: 10}, {taskName: "Task 2", value: 20}, ...]
+ *
+ * Builds and appends the y-axis
+ */
+function addYAxis(svg, yAxis, tasks) {
+  var yAxisGroup = appendYAxisGroup(svg, yAxis);
+  formatYAxisLabels(yAxisGroup, tasks);
+  addPercentageSymbol(svg);
+}
+
+function addPercentageSymbol(svg) {
+  svg
+    .append("text")
+    .text("%")
+    .attr("style", "text-anchor:end;font-weight:500;font-size:14px;")
+    .attr("x", 0 - gapSpacing)
+    .attr("dy", -3);
+}
+
+/**
+ *
+ * @param {*} yAxis
+ *
+ * Gets the tick positions of the y-axis
+ */
+function getTickPositions(yAxis) {
+  const yAxisScale = yAxis.scale();
+  const ticks = yAxisScale.ticks ? yAxisScale.ticks() : yAxisScale.domain();
+  return ticks.map((tick) => yAxisScale(tick));
+}
+
+/**
+ *
+ * @param {*} svg
+ * @param {Int} x1
+ * @param {Int} x2
+ * @param {Int} y1
+ * @param {Int} y2
+ * @param {String} strokeColor
+ * @param {Int} strokeColor
+ *
+ * Appends a horiontal line to the SVG based on custom parameters
+ */
+function addHorizontalLine(svg, x1, x2, y1, y2, strokeColor, strokeWidth) {
+  svg
+    .append("line")
+    .attr("x1", x1)
+    .attr("x2", x2)
+    .attr("y1", y1)
+    .attr("y2", y2)
+    .attr("stroke", strokeColor)
+    .attr("stroke-width", strokeWidth);
+}
+
+/**
+ *
+ * @param {*} svg
+ * @param {*} yAxis
+ * @param {Int} width
+ * @param {Obj} margin
+ *
+ * Find the postion of the Y-axis ticks and adds a horizontal
+ * line above each label and one that spans the chart area
+ */
+function addYAxisHorizontalLines(svg, yAxis, width, margin) {
+  const tickPositions = getTickPositions(yAxis);
+
+  tickPositions.forEach((posY, index) => {
+    addHorizontalLine(
+      svg,
+      -margin.left,
+      -gapSpacing,
+      posY - 2,
+      posY - 2,
+      "#D9D9D9",
+      1
+    );
+  });
+
+  addHorizontalLine(
+    svg,
+    0,
+    width + margin.right,
+    tickPositions[0] - 2,
+    tickPositions[0] - 2,
+    "#D9D9D9",
+    2
+  );
+}
+
+/**
+ *
+ * @param {*} svg
+ * @param {*} yAxis
+ *
+ * Selects the <line> elements in the x-axis and extends it
+ */
+function addXAxisTicks(svg, xAxis) {
+  d3.selectAll(".x.axis .tick line").each(function () {
+    d3.select(this)
+      .attr("y1", 3)
+      .attr("y2", 28)
+      .attr("style", "stroke-dasharray:none;");
   });
 }
 
@@ -45,227 +273,99 @@ function cleanUpChart(svg) {
 
 /**
  *
- * @param {Array} taskTypes
+ * @param {Array} YAxisLabels
+ * @param {Int} spacePerChar
+ * @param {Int} extraSpace
  *
  * Calculate the longest Y-Axis label
  */
-function calculateMaxLabelWidth(YAxisLabels) {
+function calculateYAxisWidth(YAxisLabels, spacePerChar, extraSpace = 0) {
   var YAxisLabelsCopy = YAxisLabels.slice();
   var longestLabel = YAxisLabelsCopy.sort(function (a, b) {
     return b.length - a.length;
   })[0];
-  return longestLabel.length * 7;
+  return longestLabel.length * spacePerChar + extraSpace;
 }
 
-function createChart(data, id) {
-  sortData(data);
-  var parent = d3.select(`#${id}`);
-  var parentWidth = parent.node().getBoundingClientRect().width;
-  var margin = {
-    top: 30,
-    right: 140,
-    bottom: 15,
-    left: 70,
-  };
-  var rowHeight = 40;
-  var width = parentWidth - margin.left - margin.right;
-  var height = (data.length * rowHeight);
+/**
+ *
+ * @param {String} chartSelector
+ * @param {Array} taskTypes
+ * @param {Array} tasks
+ *
+ * Builds chart using supplied selector and data
+ */
+export function createDeveloperChart(chartSelector, taskTypes, tasks) {
+  margin.left = calculateYAxisWidth(taskTypes, 15, gapSpacing);
+  var rowHeight = 36;
+  var highestValue = d3.max(tasks, (d) => d.value);
+  var height = taskTypes.length * rowHeight + margin.bottom;
+  var parent = d3.select(chartSelector);
+  var containerWidth = parent.node().getBoundingClientRect().width;
+  if (containerWidth <= 0) {
+    var closestCol = document
+      .querySelector(chartSelector)
+      .closest('[class*="col-"]');
 
-  var svg = d3
-    .select(`#${id}`)
-    .append("svg")
-    .attr("class", "developer-chart");
-
-  var g = svg.append("g");
-
-  var x = d3.scale.linear().domain([
-    0,
-    d3.max(data, function (d) {
-      return d.count;
-    }),
-  ]);
-
-  var y = d3.scale
-    .ordinal()
-    .rangeRoundBands([height, margin.top], 0.05)
-    .domain(
-      data.map(function (d) {
-        return d.label;
-      })
-    );
-
-  // Attach OS labels
-  const yAxis = d3.svg.axis().scale(y).tickSize(0).orient("left");
-  const osLabels = svg.append("g").attr("class", "OS-labels").attr("y", 0).call(yAxis);
-  // modify existing tick texts (created by D3's axis generator, hense the set timeout)
-  setTimeout(() => {
-    osLabels.selectAll(".tick text").attr("x", 0).style("text-anchor", "start");
-  }, 0);
-
-  // Attach percentage labels
-  const percentageLabels = svg
-  .append("g")
-  .attr("class", "percentage-labels")
-  .attr("transform", "translate(" + margin.left + ",0)");
-
-  // Adding the percentage values
-  percentageLabels
-    .selectAll(null)
-    .data(data)
-    .enter()
-    .append("text")
-    .attr("class", "label")
-    .attr("y", function (d) {
-      return y(d.label) + y.rangeBand() / 2 + 4;
-    })
-    .attr("x", margin.left - 10)
-    .style("text-anchor", "end")
-    .text(function (d) {
-      return `${d.count}%`;
-    });
-
-  percentageLabels.append("text")
-    .attr("class", "percentage-symbol")
-    .attr("x", margin.left + 4)
-    .attr("y", margin.top)
-    .style("text-anchor", "end")
-    .style("font-weight", "500")
-    .text('%');
-
-  // Attach lines above labels
-  const linePositions = data.map((d, i) => {
-    const yBottom =
-      i === 0 ? -y.rangeBand() / 2 : y(data[i - 1].label) + y.rangeBand();
-    const yTop = y(d.label);
-    if (i === 0) {
-      return margin.top + 3;
+    if (closestCol.clientWidth <= 0) {
+      return;
     }
-    return (yBottom + yTop) / 2;
-  });
-  svg
-    .selectAll(".line")
-    .data(linePositions)
-    .enter()
-    .append("line")
-    .attr("class", "line")
-    .attr("x1", 0)
-    .attr("x2", margin.right + 5)
-    .attr("y1", (d) => d)
-    .attr("y2", (d) => d)
-    .attr("stroke", "#ccc") // You can set the line color here
-    .attr("stroke-width", 1); // You can set the line width here
+    containerWidth = closestCol.clientWidth - margin.left;
+  }
+  var width = containerWidth - margin.right - margin.left;
 
+  var x = d3
+    .scaleLinear()
+    .domain([0, highestValue + 10])
+    .range([0, width + margin.right]);
 
-  // Attach top axis ticks
-  var topAxis = d3.svg.axis().scale(x).orient("top").tickSize(-10).tickFormat(d3.format("d"));
-  var tickValues = d3.range(0, d3.max(data, d => d.count) + 10, 10);
-  console.log(tickValues)
-  topAxis.tickValues(tickValues);
-  svg.append("g")
-    .attr("class", "top-axis")
-    .attr("transform", "translate(" + (margin.right + 40) + "," + margin.top + ")")
-    .call(topAxis);
+  var y = d3
+    .scaleBand()
+    .domain(taskTypes)
+    .rangeRound([0, height - margin.top - margin.bottom])
+    .padding(0.1);
 
-  svg.selectAll(".top-axis-line")
-  .data(tickValues)
-  .enter()
-  .append("line")
-  .attr("class", "top-axis-line")
-  .attr("x1", d => x(d) + margin.right + 40)
-  .attr("x2", d => x(d) + margin.right + 40)
-  .attr("y1", margin.top)
-  .attr("y2", margin.top + 10) // Length of the small line
-  .attr("stroke", "black") // or any color you prefer
-  .attr("stroke-width", 1);
+  var xAxis = d3.axisBottom(x).tickValues(d3.range(0, highestValue + 10, 10));
 
+  var yAxis = d3.axisRight(y).tickPadding(-margin.left).tickSize(0);
 
-  // Attach bars
-  var barsGroup = svg.append("g").attr("class", "bars");
-  var bars = barsGroup
-    .selectAll(".bar")
-    .data(data)
-    .enter()
-    .append("rect")
-    .attr("class", "bar")
-    .attr("y", function (d) {
-      return y(d.label);
-    })
-    .attr("height", y.rangeBand())
-    .attr("x", 0)
-    .attr("width", function (d) {
-      return x(d.count);
-    })
-    .attr("fill", function (d) {
-      if (d.label == "Ubuntu") {
-        return "#E95420";
-      } else {
-        return "#AEA79F";
-      }
-    });
+  sortData(tasks);
 
-  // Handle resizing
-  var update = function () {
-    svg.attr(
-      "viewBox",
-      `0 0 ${width + margin.left + margin.right} ${
-        height + margin.top + margin.bottom
-      }`
-    );
+  // Build initial chart body
+  var svg = d3
+    .select(chartSelector)
+    .append("svg")
+    .attr("class", "chart")
+    .attr("width", width + margin.left + margin.right)
+    .attr("height", height - margin.top)
+    .append("g")
+    .attr("class", "horizontal-bar-chart")
+    .attr("width", width + margin.left + margin.right)
+    .attr("height", height)
+    .attr("transform", "translate(" + margin.left + ", " + margin.top + ")");
 
-    barsGroup.attr(
-      "transform",
-      "translate(" + (margin.right - 35) + ",0)"
-    );
+  addBarsToChart(svg, tasks, x, y);
+  addXAxis(svg, height, margin, xAxis);
+  addYAxis(svg, yAxis, tasks);
 
-    g.attr("transform", "translate(" + margin.left + "," + margin.top + ")");
-
-    x.range([0, width]);
-    y.rangeRoundBands([height, margin.top], 0.1);
-
-    svg.select(".top-axis").call(topAxis);
-
-    svg
-      .selectAll(".bar")
-      .attr("x", margin.left)
-      .attr("width", function (d) {
-        return x(d.count);
-      });
-
-    osLabels.call(yAxis);
-
-    svg
-      .selectAll(".label")
-      .attr("y", function (d) {
-        return y(d.label) + y.rangeBand() / 2 + 4;
-      })
-      .attr("x", margin.left + 5);
-
-    cleanUpChart(svg);
-  };
-
-  update();
-
-  d3.select(window).on("resize", function () {
-    parentWidth = parent.node().getBoundingClientRect().width;
-    width = parentWidth - margin.left - margin.right;
-
-    update();
-  });
-
+  addYAxisHorizontalLines(svg, yAxis, width, margin);
+  addXAxisTicks(svg, xAxis);
   cleanUpChart(svg);
 }
 
 function buildCharts() {
   if (document.querySelector("#hackerearth-chart")) {
-    createChart(
-      hackerEarthData,
-      "hackerearth-chart"
+    createDeveloperChart(
+      "#hackerearth-chart",
+      hackerEarthData.labels,
+      hackerEarthData.tasks
     );
   }
   if (document.querySelector("#opensource-chart")) {
-    createChart(
-      openSourceData,
-      "opensource-chart",
+    createDeveloperChart(
+      "#opensource-chart",
+      openSourceData.labels,
+      openSourceData.tasks
     );
   }
 }
