+++ static/js/src/release-chart-old.js
@@ -285,7 +285,7 @@ function calculateYAxisWidth(YAxisLabels) {
  *
  * Builds chart using supplied selector and data
  */
-export function createChart(
+export function createSupportChartOld(
   chartSelector,
   taskTypes,
   taskStatus,
